existence: debounce on LOOKS, not cycles — doors were removed without being looked at
Asked to check whether doors are removed without the robot being near them or looking at
them. They are, and the record is unambiguous — all 12 door deaths in
etc/door_phantom_events.csv:

    fixated = 0 on TWELVE of TWELVE
    range   2.06 .. 6.73 m (median ~5.2)
    p_detect = 0.000 on five of them, central_frac = 0.00 on those same five

NOT the absence weighting, which is already correct. `ev.log_odds_delta = p_detect * d_full`
interpolates from ZERO, so a look that could not resolve the door contributes nothing and L
correctly HOLDS. The defect is one line below it, in the DEBOUNCE:

    door        if (should_remove(L)) ++streak;        // counts CYCLES
    table       streak += cycle_p_detect;              // counts LOOKS

So once L dipped below the boundary — however it got there — the streak advanced on every
subsequent cycle, INCLUDING the ones where p_detect was 0 and the channel had just, correctly,
held. The door was condemned by evidence gathered once and executed RemoveFrames cycles later,
by which time the robot was somewhere else looking at something else. `fixated = 0` on every
single death is that mechanism's signature.

★AND THE `if (integrated)` GUARD DOES NOT PREVENT IT. refrigerator and cabinet both carry a
comment saying the debounce "advances ONLY on an evidence cycle, so it counts sustained
EVIDENCE agreement (not wall-clock cycles)" — but a channel that RAN and resolved NOTHING
still sets integrated, so p_detect = 0 cycles advance it just the same. The guard was
answering a different question (did a sensor fire?) than the one that matters (could it have
seen the object?). Both are now `+= cycle_p_detect`, hoisted the same way table does it.

★REFRIGERATOR HAD A SECOND, WEAKER AUTHORITY that would have masked the fix. `plaus_remove_streak`
(the shape/singleton prior) tests the SAME L against the SAME boundary but advanced on ANY
cycle, so it would always have hit RemoveFrames first and deleted the fridge while the sensor
streak was still counting looks. It now advances only on a MEASURED, ungated cycle — which is
the rule the comment directly above it already states for its own delta ("saying it again on a
frame that observed nothing is not evidence"). Fixing only the sensor streak would have changed
almost nothing here, and I would have reported a fix that did not work.

door's two PRIOR branches (out-of-room, minimum-height) keep advancing at full weight: those
are categorical facts about where a door can be and what a door IS, deliberately independent of
visibility, and their comments argue that correctly. Only the SENSOR channel owes a look.

Also passes p_detect explicitly to integrate() in door: log_odds_delta already carried it, but
the second argument is what the frame-correlation weighting reads (rho_eff = rho*(1-p_detect)),
so defaulting it to 1.0 claimed every miss was a maximally unambiguous look. Inert while
Existence.FrameCorrelation is 0; wrong the moment anyone measures one.

All six agents rebuild clean. UNRUN.

+++ cabinet_concept/src/cabinet_instance.h
@@ -163,7 +163,7 @@ struct CabinetInstance
     // Removed only when the removal decision holds for existence_remove_frames consecutive cycles (debounce)
     // — deleting furniture warrants SUSTAINED evidence, not a transient hiccup.
     rc::exist::ExistenceBelief existence;
-    int existence_remove_streak = 0;
+    float existence_remove_streak = 0.0f;   // accumulated LOOKS (Sum p_detect), not cycles
     // Verification-gated removal (active-inference): a predicted-visible-but-absent observation from a view that
     // CANNOT resolve the cabinet (far / peripheral / edge-on) does NOT vote removal — it raises this decayed
     // go-VERIFY surprise. When it crosses existence_verify_surprise, wants_verification arms the epistemic planner
